relocate command files
