--- protostar/commands/declare/__init__.py
@@ -1,1 +1,0 @@
-from protostar.commands.declare.declare_command import DeclareCommand
