--- protostar/commands/deploy/__init__.py
@@ -1,1 +1,0 @@
-from protostar.commands.deploy.deploy_command import DeployCommand
