--- protostar/commands/format/__init__.py
@@ -1,1 +1,0 @@
-from .format_command import FormatCommand
