+++ protostar/commands/migrate_command_test.py
