--- protostar/commands/upgrade/__init__.py
@@ -1,1 +1,0 @@
-from protostar.commands.upgrade.upgrade_command import UpgradeCommand
